Updating meeting time.
Sorry for the unannounced break, I had a vacation and also got a new job which required me to move cities. Full steam ahead to complete the rest of the curriculum, all the study group members were informed in advanced of the delay, but I did not update the site.