Fix Unique Validation Errors for Lesson Completions
We have been seeing quite a few unique validation errors in New Relic.
After investigating I believe this is happening due to users clicking
the complete lesson button multiple times in a short period of time.

This disables the complete and uncomplete button until the ajax
request for the lesson completions button is completed.

This also includes a small fix for JS errors with jquery not
being found with the lesson buttons ga tracking.

That was caused by the application.js files being recentley moved
to the bottom of the page as so jquery was loaded after the script
for the ga tracking.