Change to Read Only Permission with Github Oauth
When we implemented Github oauth we must have copy and pasted the
instructions straight from the read me of the github-omniauth gem.

The example included a scope which requested full read and write
persmission to the user's profile and repos on Github. We don't need
any of that and it creeps a lot of users out when prompted to review
permissions on upon signing up.

This removes those permissions so we only request read only access to
the users profile on Github and nothing else.