Fix Error With Email and Password Sign Up's
Users who attempted to sign up with email and password were getting incorrectly
redirected to `/users` instead of to their dashboard. After investigation this was
caused by the Mail Chimp news letter sign up service.

It is currently erroring out when trying to connect to the Mail Chimp.
To stop this issue from affecting users we have wrapped the call to the
Mail Chimp api in a rescue block that will log any occurances of this error to New Relic
so we have visibility of it.