Set smtp settings after Rails is initialized
The helpers from ActionText get loaded in too early when ActionMailer is
initialized early, leading to deprecation warnings. This is a workaround
to remove the deprecation warnings. This is an issue in the main rails
repo, but the fix has not been shipped to rubygems yet.

When rails is bumped to a version with this fix, this change should be
reverted.