Change Tailwind utility class from h-screen to min-h-screen
The reason for this change is because with the h-screen class the
footer and the "sign in" section of the page overlaps on mobile
phones once phone keypad is popped up for typing. It's also very
noticeable if phone is in landscape.

Changing it from h-screen to min-h-screen class solves this issue.