Fix: Replace justify-center utility class with mx-auto and w-fit classes
The reason for this change is because of a bug where the "prev" and "next" buttons in the pagination section were getting cut off on smaller screen sizes. This was because the justify-center utility class was being used to center the "nav pagination" element, which was causing the element to overflow its container.

To fix this I used w-fit utility class which sets the width of "nav pagination" element to always fit depending on it's content. Then used mx-auto to center the element thereby fixing the bug which was introduce by using the justify-center utility class.