Chore: Move progress badge to Hotwire (#3746)
Because:
* Having the progress badge in Hotwire integrates better with the
updated complete buttons. We can change the complete button state and
update the progress badge in the same turbo stream.

This commit:
* Lazily load the course progress using Hotwire instead of using axios
to fetch it on page load.
* Update the progress when the user completes/incompletes something
using a turbo stream instead of a custom JS event.
* Extract a progress circle component from the badge which just needs to
take in a percentage - this should make it more reusable and allows us
to have a loading component when the progress is being updated.