Feature: Convert project submission flagging feature from React to Hotwire.
Because:
* We are moving our React components to Hotwire.

This commit:
* Renders Flag form within a Hotwire modal
* Amends flags controller to work with both the Hotwire and React versions of the flag form.
* Improves dark mode styling of the submission action menu dropdown.
* Add `wait_for_turbo` system test helper - lazily loaded turbo frames may not be available when Capybara tries to interact with content within it. This helper will ensures it is.

Cuprite config changes:
When working with the tests I couldn't run in non Headless mode. After some investigating, Cuprite was not being passed our custom config options set when registering the driver with Capybara.
Changing the name we use of the registered driver from "cuprite" to "odin_cuprite" fixed it. I haven't been able to find any changes documented anywhere about why this needs to be done. But it seems like using cuprite for the driver name will just use the out of the box options Cuprite has. This may also be the cause of at least some of the weird test behaviour we've been seeing on CI lately.
*