Feature: Move Dashboard project solutions to Hotwire. (#3967)
Because:
* We are replacing our React components with Hotwire.

This commit:
* Adds a title slot to project submission items - we display the lesson
title when on the dashboard and the solutions users name when on a
lesson page.
* Adds a users project submissions controller - when updating a solution
from the dashboard we need to use a user specific endpoint to ensure the
correct solution item is returned.
* Automatically like your own project when it is created - We need to
disable liking your own submissions, at least for now:
* The likes endpoint needs to return a full item component at the moment
to allow sorting to work. This conflicts with needing different item
titles on the dashboard and lesson pages. We'll be able to enable again
soon :tm:. The next phase of this project is to add a sorting feature to
the solutions list, we'll be able to update just the like counter when
that is built.
* The compromise is to automatically like your project so nobody ever
has to see their project sitting at 0 - kind of like how Reddit handles
new posts.
* A few performance improvements by preloading lesson and user solutions
when fetching solutions.