Fix: Resolves Lesson Preview share bug on Safari. (#4032)
## Because
This PR resolves the bug that @ManonLef reported with sharing lesson
previews. The issue lives
[here](https://github.com/TheOdinProject/theodinproject/issues/4014).

The Clipboard's API's usage across browser's is very inconsistent, which
led to some of the changes in this PR.

## This PR
- Adds a method to detect the browser in the Share Button component.
- Uses this method to determine how to write the preview link to the
clipboard.
- If the browser is Safari, the ClipboardItem object is used.
- If the browser is anything else, the existing clipboard logic fires.

## Issue
Closes #4014 

## Additional Information
I'm not in love with having to detect the Browser the way I did, but
felt this was the simplest solution for now. I found no documentation on
using browser features as a more-sturdy approach.

## Pull Request Requirements
- [x] I have thoroughly read and understand [The Odin Project
Contributing
Guide](https://github.com/TheOdinProject/theodinproject/blob/main/CONTRIBUTING.md)
- [x] The title of this PR follows the `keyword: brief description of
change` format, using one of the following keywords:
-   [x] The `Because` section summarizes the reason for this PR
- [x] The `This PR` section has a bullet point list describing the
changes in this PR
- [x] I have verified all tests and linters pass after making these
changes.
- [x] If this PR addresses an open issue, it is linked in the `Issue`
section
-   [x] If applicable, this PR includes new or updated automated tests