Feature: add `report lesson` and `how to contribute` links in lesson footers (#4057)
<!-- Thank you for taking the time to contribute to The Odin Project. In
order to get this pull request (PR) merged in a reasonable amount of
time, you must complete this entire template. -->

## Because
<!-- Summarize the purpose or reasons for this PR, e.g. what problem it
solves or what benefit it provides. -->
Add links to lessons for learners to report issues with lessons instead
of getting taken to make the changes themselves.

## This PR
<!-- A bullet point list of one or more items describing the specific
changes. -->

- Adds a `Report an issue with this lesson` link next to the `Improve
this lesson` link at the bottom of every lesson
- Adds a `Learn how to contribute` link as well.
- Adds a font awesome flag icon and a book icon for the respective links
- Adds a small amount of padding between the links
- **Moves links above the completion component**
- **Adds responsiveness to show only icon with tooltip on mobile view**

## Issue
<!--
If this PR closes an open issue in this repo, replace the XXXXX below
with the issue number, e.g. Closes #2013.

If this PR closes an open issue in another TOP repo, replace the #XXXXX
with the URL of the issue, e.g. Closes
https://github.com/TheOdinProject/curriculum/issues/XXXXX

If this PR does not close, but is related to another issue or PR, you
can link it as above without the 'Closes' keyword, e.g. 'Related to
#2013'.
-->


## Additional Information
<!-- Any other information about this PR, such as a link to a Discord
discussion. -->

~This is a draft PR. More work needs to be done on where the report an
issue will be linked to.~ ready for review.

discord link -
https://discord.com/channels/505093832157691914/991035463949254767/1135930665502769226

## Pull Request Requirements
<!-- Replace the whitespace between the square brackets with an 'x',
e.g. [x]. After you create the PR, they will become checkboxes that you
can click on. -->
- [x] I have thoroughly read and understand [The Odin Project
Contributing
Guide](https://github.com/TheOdinProject/theodinproject/blob/main/CONTRIBUTING.md)
- [x] The title of this PR follows the `keyword: brief description of
change` format, using one of the following keywords:
  - `Feature` - adds new or amends existing user-facing behavior
- `Chore` - changes that have no user-facing value, refactors,
dependency bumps, etc
  - `Fix` - bug fixes
-   [x] The `Because` section summarizes the reason for this PR
- [x] The `This PR` section has a bullet point list describing the
changes in this PR
- [ ] I have verified all tests and linters pass after making these
changes.
- [ ] If this PR addresses an open issue, it is linked in the `Issue`
section
-   [ ] If applicable, this PR includes new or updated automated tests