How to ask a technical question: Fix a broken link (#4046)
<!-- Thank you for taking the time to contribute to The Odin Project. In
order to get this pull request (PR) merged in a reasonable amount of
time, you must complete this entire template. -->

## Because
<!-- Summarize the purpose or reasons for this PR, e.g. what problem it
solves or what benefit it provides. -->
- The `improve this lesson on github` link was broken 

## This PR
<!-- A bullet point list of one or more items describing the specific
changes. -->
- Fix the broken link

## Issue
<!--
If this PR closes an open issue in this repo, replace the XXXXX below
with the issue number, e.g. Closes #2013.

If this PR closes an open issue in another TOP repo, replace the #XXXXX
with the URL of the issue, e.g. Closes
https://github.com/TheOdinProject/curriculum/issues/XXXXX

If this PR does not close, but is related to another issue or PR, you
can link it as above without the 'Closes' keyword, e.g. 'Related to
#2013'.
-->

#4033
## Additional Information
<!-- Any other information about this PR, such as a link to a Discord
discussion. -->
I didn't realize this issue was something in regard to the
`theodinproject` repository and not the `curriculum` repository of TOP.
This is my first time contributing to this one, and I'm not quite sure
if I solved the issue correctly. Below is the corrected link.

- The current broken link is
https://github.com/TheOdinProject/theodinproject/edit/main/app/views/static_pages/how_to_ask.html.erb
- The correct link should be
https://github.com/TheOdinProject/theodinproject/edit/main/app/views/guides/community/how_to_ask.html.erb

## Pull Request Requirements
<!-- Replace the whitespace between the square brackets with an 'x',
e.g. [x]. After you create the PR, they will become checkboxes that you
can click on. -->
- [x] I have thoroughly read and understand [The Odin Project
Contributing
Guide](https://github.com/TheOdinProject/theodinproject/blob/main/CONTRIBUTING.md)
- [x] The title of this PR follows the `keyword: brief description of
change` format, using one of the following keywords:
  - `Feature` - adds new or amends existing user-facing behavior
- `Chore` - changes that have no user-facing value, refactors,
dependency bumps, etc
  - `Fix` - bug fixes
-   [x] The `Because` section summarizes the reason for this PR
- [x] The `This PR` section has a bullet point list describing the
changes in this PR
- [ ] I have verified all tests and linters pass after making these
changes.
- [x] If this PR addresses an open issue, it is linked in the `Issue`
section
-   [ ] If applicable, this PR includes new or updated automated tests

---------

Co-authored-by: Kevin <kevinmulhern23@gmail.com>