Feature: Add screencast link to README to giving overview for contributors (#4090)
<!-- Thank you for taking the time to contribute to The Odin Project. In
order to get this pull request (PR) merged in a reasonable amount of
time, you must complete this entire template. -->

## Because
<!-- Summarize the purpose or reasons for this PR, e.g. what problem it
solves or what benefit it provides. -->

This PR adds an interactive screencast to the README that gives a
general overview of the contributing guide to new collaborators. It
helps people who may be new to contributing to this or any other repo
get a sense of the code base and how to get it running locally. I used a
new Scrimba feature that lets one make a video screencast with all the
assets of a github repo.

## This PR
<!-- A bullet point list of one or more items describing the specific
changes. -->

- Added link to interactive screencast (scrim) to __README__ in the
__Contributing__ section

## Issue
<!--
If this PR closes an open issue in this repo, replace the XXXXX below
with the issue number, e.g. Closes #2013.

If this PR closes an open issue in another TOP repo, replace the #XXXXX
with the URL of the issue, e.g. Closes
https://github.com/TheOdinProject/curriculum/issues/XXXXX

If this PR does not close, but is related to another issue or PR, you
can link it as above without the 'Closes' keyword, e.g. 'Related to
#2013'.
-->
Related to #3470

## Additional Information
<!-- Any other information about this PR, such as a link to a Discord
discussion. -->

Here is a link to the screencast (scrim):
https://scrimba.com/scrim/cEggQncE
I can also turn this into an image that links to the scrim video instead
of text if you prefer that.

## Pull Request Requirements
<!-- Replace the whitespace between the square brackets with an 'x',
e.g. [x]. After you create the PR, they will become checkboxes that you
can click on. -->
- [x] I have thoroughly read and understand [The Odin Project
Contributing
Guide](https://github.com/TheOdinProject/theodinproject/blob/main/CONTRIBUTING.md)
- [x] The title of this PR follows the `keyword: brief description of
change` format, using one of the following keywords:
  - `Feature` - adds new or amends existing user-facing behavior
- `Chore` - changes that have no user-facing value, refactors,
dependency bumps, etc
  - `Fix` - bug fixes
-   [x] The `Because` section summarizes the reason for this PR
- [x] The `This PR` section has a bullet point list describing the
changes in this PR
- [ ] I have verified all tests and linters pass after making these
changes.
- [x] If this PR addresses an open issue, it is linked in the `Issue`
section
-   [ ] If applicable, this PR includes new or updated automated tests