Chore: curriculum task spec failures (#4240)
<!-- Thank you for taking the time to contribute to The Odin Project. In
order to get this pull request (PR) merged in a reasonable amount of
time, you must complete this entire template. -->

## Because
<!-- Summarize the purpose or reasons for this PR, e.g. what problem it
solves or what benefit it provides. -->
I often run the test suite excluding system specs (because I like to go
fast), but I noticed that the test for the curriculum update task always
fails when run this way. It also fails when run alone, which you should
be able to verify by quickly running:
```
bundle exec rspec spec/tasks/curriculum_spec.rb
```
This seems to have emerged from #3893, which added precompiling assets
when the test run includes system specs.

This test *does* work when the full suite is run, but it felt not great
to me that it's dependent on the asset precompile happening.

## This PR
<!-- A bullet point list of one or more items describing the specific
changes. -->
- Adds the usual rake task spec boilerplate to curriculum task spec. The
absence of this is what causes it to fail when run without the asset
precompile
- The above change gets the test passing when run without system specs
(good) but now they fail when run with system specs (not so good). Kevin
observes this here in the aforementioned [here in the aforementioned
PR](https://github.com/TheOdinProject/theodinproject/pull/3893#discussion_r1244426518).
Something about the task loading is interfering with this test. So
instead of loading the tasks into the Rails app to invoke
`assets:precompile`, I use `system` to run it as a shell command. All
tests are green and assets still precompile for runs that include system
specs. This should also still be fine on CI? Definitely let me know if
there's some mistaken assumptions with that.
- Renames `spec/support/precomple_assets.rb` to
`spec/support/precompile_assets.rb` (minor incidental typo correction)

## Issue
<!--
If this PR closes an open issue in this repo, replace the XXXXX below
with the issue number, e.g. Closes #2013.

If this PR closes an open issue in another TOP repo, replace the #XXXXX
with the URL of the issue, e.g. Closes
https://github.com/TheOdinProject/curriculum/issues/XXXXX

If this PR does not close, but is related to another issue or PR, you
can link it as above without the 'Closes' keyword, e.g. 'Related to
#2013'.
-->
Closes #XXXXX

## Additional Information
<!-- Any other information about this PR, such as a link to a Discord
discussion. -->


## Pull Request Requirements
<!-- Replace the whitespace between the square brackets with an 'x',
e.g. [x]. After you create the PR, they will become checkboxes that you
can click on. -->
- [x] I have thoroughly read and understand [The Odin Project
Contributing
Guide](https://github.com/TheOdinProject/theodinproject/blob/main/CONTRIBUTING.md)
- [x] The title of this PR follows the `keyword: brief description of
change` format, using one of the following keywords:
    - `Feature` - adds new or amends existing user-facing behavior
- `Chore` - changes that have no user-facing value, refactors,
dependency bumps, etc
    - `Fix` - bug fixes
-   [x] The `Because` section summarizes the reason for this PR
- [x] The `This PR` section has a bullet point list describing the
changes in this PR
- [x] I have verified all tests and linters pass after making these
changes.
- [x] If this PR addresses an open issue, it is linked in the `Issue`
section
-   [x] If applicable, this PR includes new or updated automated tests