Feature: Add support for parsing backtick fenced codeblocks (#4241)
<!-- Thank you for taking the time to contribute to The Odin Project. In
order to get this pull request (PR) merged in a reasonable amount of
time, you must complete this entire template. -->

## Because
<!-- Summarize the purpose or reasons for this PR, e.g. what problem it
solves or what benefit it provides. -->
Kramdown's default supported syntax for delimiting codeblocks (fencing
with three `~`) is pretty unconventional. This especially causes
problems with VSCode's markdown editing, which has strong (negative)
opinions of the tilde fenced codeblocks. This makes contributing more
annoying because you have to pay attention to this quirk. It also makes
maintenance more annoying because contributors may accidentally
autoformat the tildes to backticks.

## This PR
<!-- A bullet point list of one or more items describing the specific
changes. -->
- Add [kramdown-parser-gfm](https://github.com/kramdown/parser-gfm) gem
to support "Github Flavored Markdown". One of the key features of this
is support for backtick fenced codeblocks
- Edit the `MarkdownConverter` so that documents are parsed using
github-flavored markdown

## Issue
<!--
If this PR closes an open issue in this repo, replace the XXXXX below
with the issue number, e.g. Closes #2013.

If this PR closes an open issue in another TOP repo, replace the #XXXXX
with the URL of the issue, e.g. Closes
https://github.com/TheOdinProject/curriculum/issues/XXXXX

If this PR does not close, but is related to another issue or PR, you
can link it as above without the 'Closes' keyword, e.g. 'Related to
#2013'.
-->
Closes #XXXXX

## Additional Information
<!-- Any other information about this PR, such as a link to a Discord
discussion. -->
I've been unable to find any problems or new oddities caused by parsing
with GFM. I reset my db and re-imported content locally using it and
have been unable to find anything strange happening. But also I'm only
one person and can't realistically browse everything. It's worth noting
that the tilde syntax is still valid and will be correctly parsed, which
is convenient for our usecase.

There are arguably some drawbacks to this. It's an extra dependency, and
it's one that doesn't seem to get much attention. We could potentially
look at rolling a custom solution for this, especially since the
backtick codeblocks are the only thing I think we really care about.

This might be as simple as finding a way to modify [Kramdown's regex
matching on
codeblocks](https://github.com/gettalong/kramdown/blob/master/lib/kramdown/parser/kramdown/codeblock.rb#L33-L34).

Also if this is accepted, we probably want to spin up a bunch of tickets
for converting the curriculum markdown documents over to using backtick
fenced codeblocks. And also would need to update the style guide to tell
contributors to use the backtick format.

## Pull Request Requirements
<!-- Replace the whitespace between the square brackets with an 'x',
e.g. [x]. After you create the PR, they will become checkboxes that you
can click on. -->
- [x] I have thoroughly read and understand [The Odin Project
Contributing
Guide](https://github.com/TheOdinProject/theodinproject/blob/main/CONTRIBUTING.md)
- [x] The title of this PR follows the `keyword: brief description of
change` format, using one of the following keywords:
    - `Feature` - adds new or amends existing user-facing behavior
- `Chore` - changes that have no user-facing value, refactors,
dependency bumps, etc
    - `Fix` - bug fixes
-   [x] The `Because` section summarizes the reason for this PR
- [x] The `This PR` section has a bullet point list describing the
changes in this PR
- [x] I have verified all tests and linters pass after making these
changes.
- [x] If this PR addresses an open issue, it is linked in the `Issue`
section
-   [x] If applicable, this PR includes new or updated automated tests