Chore: Remove Project Submission Title Component (#4388)
Because:
- We previously used the ProjectSubmissions::ItemComponent for the
current user and other users' project submissions. This required us to
make the title dynamic and configurable from outside the component. But
we have since introduced a separate component for the current users
solution and can now simplify the project submission item component by
hard coding the title directly in the view.