Chore: Only Update Likes Component When Using Likes (#4395)
Because:
- Previously we were replacing the entire solutions component when
liking/unliking. We did that to allow the list to auto sort based on the
likes count of each solution. But we have recently replaced auto sorting
with a user controlled manual sort. This allows us to be more efficient
with likes, and only send the HTML we need over the wire.

This commit:
* Update the likes component instead of the solutions component with
Turbo when liking/unliking
* Removes the likes/create.turbo_stream.html view. This must have
accidentally got into the codebase at some point, we only use an update
action with likes.
* Updates the like icon tool tip to show an appropriate message based on
the state - before it was only showing "Like submission"