Fix: Script tags could execute in from previews
Because:
- We weren't sanitizing HTML within the outputted preview.

This commit:
- Run the output through Rails's [sanitize](https://api.rubyonrails.org/classes/ActionView/Helpers/SanitizeHelper.html#method-i-sanitize) helper.

Notes
I tried a few different approaches with this:

Scrubbing the output before storing it proved to be very difficult. We're storing the content in markdown which makes it hard to differentiate legitimate script elements that will be displayed within code examples vs top level malicious script elements.

But, scrubbing the output before rendering should be enough to mitigate this issue.