Feature: Add link to lesson commit history (#4570)
<!-- Thank you for taking the time to contribute to The Odin Project. In
order to get this pull request (PR) merged in a reasonable amount of
time, you must complete this entire template. -->

## Because
<!-- Summarize the purpose or reasons for this PR, e.g. what problem it
solves or what benefit it provides. -->
It would be good to see when a lesson was last updated and go directly
to the lesson's commit history to see the changelog, directly from the
lesson page itself. This should help alleviate some confusion when
something in the lesson changes when someone is midway through - they'll
now be able to see that it was changed and find out what the change was,
without having to go to GitHub of their own accord and navigate through
the files to find out.

## This PR
<!-- A bullet point list of one or more items describing the specific
changes. -->
- Adds a github commit url helper method
- Adds a link to the GitHub lesson links at the end of a lesson page,
showing the time of last update, and linking to that lesson's commit
history.


## Issue
<!--
If this PR closes an open issue in this repo, replace the XXXXX below
with the issue number, e.g. Closes #2013.

If this PR closes an open issue in another TOP repo, replace the #XXXXX
with the URL of the issue, e.g. Closes
https://github.com/TheOdinProject/curriculum/issues/XXXXX

If this PR does not close, but is related to another issue or PR, you
can link it as above without the 'Closes' keyword, e.g. 'Related to
#2013'.
-->
Closes #XXXXX

## Additional Information


![image](https://github.com/TheOdinProject/theodinproject/assets/122839503/bf03640b-f751-4816-a053-f3ee09e07b8d)



## Pull Request Requirements
<!-- Replace the whitespace between the square brackets with an 'x',
e.g. [x]. After you create the PR, they will become checkboxes that you
can click on. -->
- [x] I have thoroughly read and understand [The Odin Project
Contributing
Guide](https://github.com/TheOdinProject/theodinproject/blob/main/CONTRIBUTING.md)
- [x] The title of this PR follows the `keyword: brief description of
change` format, using one of the following keywords:
    - `Feature` - adds new or amends existing user-facing behavior
- `Chore` - changes that have no user-facing value, refactors,
dependency bumps, etc
    - `Fix` - bug fixes
-   [x] The `Because` section summarizes the reason for this PR
- [x] The `This PR` section has a bullet point list describing the
changes in this PR
- [x] I have verified all tests and linters pass after making these
changes.
- [ ] If this PR addresses an open issue, it is linked in the `Issue`
section
-   [ ] If applicable, this PR includes new or updated automated tests