Chore: Abort lesson import when bad credentials are provided (#4768)
Because:
- If your GitHub Api token is incorrect, the lesson import task will
silently fail. It will run through every lesson, incrementing the
progress bar, and look as if it has succeeded. Resulting in a lot of
confusion when you go to a lesson page and the content is blank.

This commit:
- Abort the lesson import task if an incorrect GitHub API token is
provided and print out a message about what might have gone wrong.